Feat: Match javascript locale to Rails locale
Some plugins use methods other than, or in addition to, the user-selected locale. This is a simple change that passes the current Javascript I18n.locale along with the Ajax request to ensure that the backend locale matches the front end when necessary. This should be a non-breaking change, and most people will see no behavior change whatsoever.